tars-ai/api/requirements.txt
Adafruit_PCA9685==1.0.1
Flask==3.0.3
Flask_Cors==5.0.0
langchain==0.2.16
langchain_community==0.2.16
langchain_openai==0.1.23
python-dotenv==1.0.1
Requests==2.32.3
chromadb==0.4.1
